chore: give the studio its own release, and prune what it does not need
Stage 10.

release-studio.yml publishes to Frostn1/frost-studio rather than here, and that
split is the whole point: `releases/latest` resolves to one release per
*repository*, and the manager's updater endpoint — baked into every shipped
binary — points at this one. A studio release published here would hand every
manager install a studio build as its next update. Separate repo, separate
`releases/latest`, separate signing key, so a leaked studio token cannot push a
manager update either.

Tags are `studio-v0.1.0` here and land as `v0.1.0` over there. Checked that
`studio-v0.1.0` does not match release.yml's `v*` glob, so the two workflows
cannot both fire.

The prune came out as the plan expected, and no further:

- `trash` is gone from the studio. Freeing it meant lifting library's Trash block
  into a manager-side `trashbin.rs` — on macOS the `trash` crate drives Finder
  over AppleScript, which is why that code calls NSFileManager directly and drags
  `objc2` in to do it. None of that belongs in a binary that only paints.
- `uuid` is behind a `mint-install-id` feature only the manager enables. A build
  without it never mints an install id, and an empty id means "do not report" —
  the right answer for a second app sharing one config, since the id belongs to
  the install and the manager is what created it.
- `objc2` and `uuid` still appear in the studio's tree, but transitively:
  tauri-plugin-dialog → rfd → dispatch2 → block2, and tauri-utils → infer → cfb.
  Not ours to remove, and worth writing down so the next person does not go
  looking.

Everything else the plan listed — the voice stack, unrar/sevenz/mega, scraper,
blowfish, ed25519-dalek, memory-stats, obfstr, tungstenite — was never in the
studio's tree: its manifest was written from what it needs rather than trimmed
from the manager's.

Verified: cargo check clean on host and x86_64-pc-windows-gnu; 913 + 362 + 213
tests pass; census green; typecheck, lint and both builds clean.
