+++ .github/workflows/release-studio.yml
@@ -1,0 +1,124 @@
+name: Release Frost's Studio
+
+# Build and publish Frost's Studio.
+#
+# Source lives in this repo; releases go to Frostn1/frost-studio. That split is not
+# cosmetic: `releases/latest` resolves to one release per *repository*, and the manager's
+# updater endpoint — baked into every shipped binary since v0.1 — points at this repo's.
+# Publishing the studio here would hand every manager install a studio build as its next
+# update. So the studio gets its own repo, its own `releases/latest`, and its own signing
+# key: with both separate, a leaked studio token still cannot push a manager update.
+#
+# Tags are `studio-v0.1.0` here and land as `v0.1.0` over there, so neither repo's tag
+# namespace collides. `release.yml` triggers on `v*`, which does not match `studio-v*`.
+
+on:
+  push:
+    tags:
+      - "studio-v*"
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - platform: windows-latest
+            args: ""
+          - platform: macos-latest
+            args: "--target aarch64-apple-darwin"
+          # Pinned to 22.04 for the same reason the manager's is: the AppImage inherits the
+          # builder's glibc as its floor.
+          - platform: ubuntu-22.04
+            args: ""
+
+    runs-on: ${{ matrix.platform }}
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: npm
+
+      - name: Setup Rust
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: ${{ matrix.platform == 'macos-latest' && 'aarch64-apple-darwin' || '' }}
+
+      # No libasound2-dev here: the voice stack is the manager's, and the studio does not
+      # build cpal. Everything else mirrors the manager's list.
+      - name: Install Linux system dependencies
+        if: startsWith(matrix.platform, 'ubuntu')
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libwebkit2gtk-4.1-dev \
+            libgtk-3-dev \
+            libayatana-appindicator3-dev \
+            librsvg2-dev \
+            libsoup-3.0-dev \
+            patchelf \
+            squashfs-tools
+
+      - name: Cache Rust build
+        uses: Swatinem/rust-cache@v2
+        with:
+          workspaces: . -> target
+
+      - name: Install dependencies
+        run: npm ci
+
+      # The studio's Protect tab needs the same private module the manager does; without it
+      # the tab hides itself rather than failing, so a fork still builds.
+      - name: Fetch private sidecar module
+        if: ${{ github.repository == 'Frostn1/mxb-app' }}
+        uses: actions/checkout@v4
+        with:
+          repository: Frostn1/mxbapp-private
+          ssh-key: ${{ secrets.SIDECAR_DEPLOY_KEY }}
+          path: .sidecar-src
+
+      - name: Apply sidecar module
+        if: ${{ github.repository == 'Frostn1/mxb-app' }}
+        shell: bash
+        run: bash .sidecar-src/sync.sh "$GITHUB_WORKSPACE"
+
+      - name: Tag without the studio- prefix
+        id: tag
+        shell: bash
+        run: |
+          raw="${{ github.ref_type == 'tag' && github.ref_name || format('studio-v0.0.{0}', github.run_number) }}"
+          echo "name=${raw#studio-}" >> "$GITHUB_OUTPUT"
+
+      - name: Build and publish release
+        uses: tauri-apps/tauri-action@v0
+        env:
+          # A fine-grained PAT with contents:write on frost-studio only. Deliberately not
+          # GITHUB_TOKEN: that cannot write to another repository.
+          GITHUB_TOKEN: ${{ secrets.STUDIO_RELEASE_TOKEN }}
+          VITE_APP_NAME: "Frost's Studio"
+          TAURI_SIGNING_PRIVATE_KEY: ${{ secrets.STUDIO_SIGNING_PRIVATE_KEY }}
+          TAURI_SIGNING_PRIVATE_KEY_PASSWORD: ${{ secrets.STUDIO_SIGNING_PRIVATE_KEY_PASSWORD }}
+        with:
+          owner: Frostn1
+          repo: frost-studio
+          # Required when publishing across repositories: the action has no commit of its
+          # own over there to hang the release off.
+          releaseCommitish: main
+          projectPath: apps/studio
+          tagName: ${{ steps.tag.outputs.name }}
+          releaseName: "Frost's Studio ${{ steps.tag.outputs.name }}"
+          # Named here rather than renamed afterwards: this applies before latest.json is
+          # written, so the manifest and the assets cannot disagree.
+          releaseAssetNamePattern: "Frost-Studio-[version]-[arch][ext]"
+          releaseDraft: false
+          prerelease: ${{ github.ref_type != 'tag' || contains(github.ref_name, '-beta') || contains(github.ref_name, '-rc') }}
+          args: ${{ matrix.args }}
