+++ apps/manager/src-tauri/Cargo.toml
@@ -16,7 +16,7 @@ tauri-build = { version = "2", features = [] }
 [dependencies]
 # The shared core: the game's file formats, re-exported at this crate's root so the
 # `crate::edf::…` call sites throughout this app keep resolving.
-mxb-core = { path = "../../../crates/core" }
+mxb-core = { path = "../../../crates/core", features = ["mint-install-id"] }
 # Secure-content packing (mxbsecure). Public crates, always compiled; the format and its use
 # live in the gitignored `src/mxbsecure.rs` sidecar, gated by cfg(mxbsecure).
 chacha20poly1305 = "0.10"
